+++ es/unit.go
@@ -19,7 +19,7 @@ type Unit interface {
 	Commit(ctx context.Context) error
 	Load(ctx context.Context, id string, aggregateName string, out interface{}) error
 	Find(ctx context.Context, aggregateName string, filter filters.Filter, out interface{}) error
-	Count(ctx context.Context, aggregateName string, filter filters.Filter) (int64, error)
+	Count(ctx context.Context, aggregateName string, filter filters.Filter) (int, error)
 }
 
 type unit struct {
@@ -80,7 +80,7 @@ func (u *unit) Find(ctx context.Context, aggregateName string, filter filters.Fi
 	return u.tx.Find(ctx, u.serviceName, aggregateName, namespace, filter, out)
 }
 
-func (u *unit) Count(ctx context.Context, aggregateName string, filter filters.Filter) (int64, error) {
+func (u *unit) Count(ctx context.Context, aggregateName string, filter filters.Filter) (int, error) {
 	namespace := NamespaceFromContext(ctx)
 	return u.tx.Count(ctx, u.serviceName, aggregateName, namespace, filter)
 }
